qml: use background coloring in grid expansion image
This is because if the image is translucent, it looks
off with the shadow in the background as the shadow
gets exposed.

We already use background coloring in most other
delegates, so this should make the expansion image
aligned with the rest.

It requires further investigation whether we should
switch to `ImageExt` in all cases. Historically we
are not using it in various places, such as playlist
delegate because rounding has not been needed there
and `ImageExt` (formerly `RoundImage`) used to be
used only for rounding.

Now, it is much more than mere rounding, and can be
useful in other cases, such as having preserve aspect
crop fill mode without requiring clipping.

Although `ImageExt` can be considered lightweight,
it is still heavier than bare `Image`, so switching
from `Image` to `ImageExt` needs to be justified.
Switching just because of background coloring may
not be a proper justification, because the same can
be achieved with a `Rectangle`, provided that the
image is not rounded.