contrib: gpg-error: fix --disable-threads usage
The locking API is still called with Win32/POSIX structures even though the locks
are not defined the same way with --disable-threads.
We just turn the locking code off in --disable-threads (locking does nothing).

Fixes a crash on exit on Windows where _gpgrt_fflush (NULL) is always called,
locking estream_list_lock that is not initialized as expected.