qml: do not use layering when background is opaque in `FadingEdge`
This is an optimization case for opaque background color.

With custom blending, we could potentially achieve the same
even if the background is not opaque, but that requires more
investigation since in such a case, not only the view elements
would be affected but the whole primitives drawn before in the
same area, while the fading edge effect should only be applied
to the primitives of the view (and not the background).