qml: do not track current index change in playlist list view
Current index of the view and the current index of the
playlist are not a one-to-one match.

The view, when applicable, may track the playlist
current index. However, when the view current index
changes, such as due to keyboard navigation, it is
fair that the view current index deviates from the
playlist current index. It is also fair to not
change the current index of the view at all, but
rather make the new playlist current item visible,
which is what is the case with this patch.

This fixes the situation of the view scrolling
unwantedly when, for example, playlist current index
changes due to adding or removing items preceding to
the playlist current item while the current item
remains the same.