qml: get rid of the opacity behavior/animation in `ScrollBarExt`
The important animation here is the size animation, opacity
animation is something that is not really important.

We do not want to animate the opacity at initialization. In
other cases, we often enable the behavior when the component
completes, so that at initialization opacity change is not
animated. Since the animation is not important, we can just
get rid of it instead.