qml: do not delay implicit height of footer drag accessories anymore
The binding was made delayed to fix a binding loop, which was
not symptomatic. This delay causes the scroll bar to appear
intermittently.

I no longer get the warning without the delay, so it seems
safe to get rid of the delay. Besides, there seems no apparent
reason to have a binding loop here, footer's (y) position
should not depend on footer's height to cause a binding loop.