qt: get rid of `QSGRoundedRectangularImageNode`
There has been nothing that used this node since we started
to use SDF to draw rounded images (40dcb6a4).

Using this node would be beneficial where having a custom
geometry with triangle strips is more favorable than
having an expensive fragment shader. At the same time, if
MSAA is enabled, there would be anti-aliasing while keeping
the primitive opaque (which is important for effective
batching), where with the SDF approach this is not possible.

Currently we do not offer a way to enable MSAA, and without
MSAA, it lacked AA completely. Implementing AA with SDFs
is much easier than implementing "vertex" AA with custom
geometry.

Requesting MSAA by default is still on the table though,
because that way built-in Qt items would also not need to
use "vertex" AA, and as Qt docs noted, this would improve
batching overall. This is mostly relevant for round items,
such as `Rectangle` with `radius`.