qml: fix binding issues in table view with Qt 6.4
passing binding to object to createObject fails with Qt6.4, see QTBUG-125095.
Using `var` properties as suggested in the Qt bug report isn't enough to
workaround the issue.

The proposed approach is to create a local QtObject containing the bindings and
pass it in the delegate creation (without binding), then we create readonly
properties to the model properties for convenience

fix: #29430