macosx: PLItem: clamp pos when inserting
Pos needs to be clamped to the amount of items in the array, else
we might try to insert at an invalid position in the array.

This can happen because the playlist can change till we process the
insertion request, so an item that would have been inserted before at
position 0 is dropped and then we try to insert the next item at
position 1 which would lead to an exception as it is outside the array
bounds.

This is easy to reproduce when using the flattened playlist and dragging
a folder with a few items and a subfolder onto the icon or playlist.

Fix #28660