qt: do not signal current item change unnecessarily in `PlaylistController` #2
This was missed in 776121f3 because I thought, as the
callback name implied, it would be called when the current
item changed so checking if the current item changed
would be redundant. However, the callback is actually
for current index change, so we also have to compare
the current item there before signalling the change
to prevent unnecessarily signalling the change.