input: add previous-frame support
Add `decoder_prevframe.c` that handle most of the previous-frame logic:
trigger a seek to `current_frame - n` (n is 1 at the start).

Store 1 decoded frame until we reach the current frame. When the
previous frame is found, return it and trigger another seek when needed.
In case the seek missed the current frame, increase n and trigger
another seek.

This solution doesn't require more memory, it is able to play backward
"in real-time" (when pressing the previous-frame key) most of 1080p
videos.

It is also able to play 4K 60FPS videos with few slow-downs (when seeking).

Fixes #10252