tools: allow newer major versions
When the user has CMake 4.x we should not build CMake 3.x
because we require at least 3.21.0.

This matching major version was introduced for protoc in ee5fad6a354d4fde1bb98f715e0455c603599554. But
it has moved to contribs and other tools (not matching a library)
should be backward compatible.

If there is another tool that needs to match the major version exactly,
we should at a special check for that.