qml: do not adjust fill mode depending on graphics backend in `AudioGridItem`
`ImageExt` uses a clip node if backend is not rhi, which is not
a concern for non-rhi (software or openvg) backend.

Batch rendering is only a thing with the rhi backend, and non-
rhi mode is already not meant for having high performance.

This fixes preserve aspect crop not used with software mode in
audio grid delegate.