win32: dirs: do not create some system folders on behalf of the user
These should not be created by VLC
* VLC_HOME_DIR (%userprofile%\Documents)
* VLC_MUSIC_DIR (%userprofile%\Music)
* VLC_PICTURES_DIR (%userprofile%\Pictures)
* VLC_VIDEOS_DIR (%userprofile%\Videos)
* VLC_DESKTOP_DIR = VLC_HOME_DIR (%userprofile%\Documents)
* VLC_TEMPLATES_DIR = VLC_HOME_DIR (%userprofile%\Documents)
* VLC_PUBLICSHARE_DIR = VLC_HOME_DIR (%userprofile%\Documents)
* VLC_DOCUMENTS_DIR = VLC_HOME_DIR (%userprofile%\Documents)
* VLC_DOWNLOAD_DIR = VLC_HOME_DIR (%userprofile%\Documents)

VLC_VIDEOS_DIR/VLC_MUSIC_DIR/VLC_DOWNLOAD_DIR are used by the es_out.
But if the system doesn't have them or if they are not usable we should not write in them.
The user has the option to set "input-record-path" to set a usable folder.

VLC_PICTURES_DIR is used to save the snapshots and there is "snapshot-path"
to use a different folder.

These should be created by VLC if needed:
* VLC_CONFIG_DIR
* VLC_DATA_DIR
* VLC_CACHE_DIR

But they are not actually created during the config_GetAppDir() call.

Fixes #29488