qml: disable clipping if there is depth buffer with csd shadows
We can make use of the depth buffer here, which is better than
clipping because we don't need to spend effort on clipping
anything since the content pixels will not be painted in the
shadow area as the shadow pixels obscure the content pixels.

On top of that, if early-z is applicable, the fragment shader
for the content pixels that are obscured by the quasi-opaque
shadow will not be executed.