Revert "qt: call `vlc_window_ReportSize()` before the video surface changes are committed"
This reverts commit a8cc3f6db15da074287fcd6cb9fdb7d938fff971.

Since report size is asynchronous, the idea was to report the size preemptively to not
wait for the compositor to actually adjust the video window so that the video window
would not be adjusted too early, particularly before the video is adjusted. However,
this turned out to be problematic with HWND window type.

This approach is also technically incorrect, because report size expects the current
size of the window and not the pending size. Preemptive reporting violates such
expectation.