decoder: accurately count frames_countdown
This will allow the frames_countdown variable to match with the actual
number of frame-next requests.

Indeed, this variable was incremented, regardless of the number of
pictures ready to be displayed by the vout (just decoded).

This was resulting in a buffering gap: If there were 20 frames on the
paused vout, the same amount of pictures was kept ahead for all future
frame-next requests because of the frames_countdown increase. Because of
that, it was also impossible to know when we reached EOF on next-frame.

Furthermore, this commit is needed for the new and future next-frame
buffering workaround as it allows finer buffering (ask buffering only
when needed, and not 20 frames in advance).

Refs #29487