mmdevice: remove lock from OnDefaultDeviceChange
sys->device_name may change after we read it.

If we trigger an unwanted restart (read as NULL and became a user selected value),
it's wasteful but supported. The new non-NULL sys->device_name will be used during
both restarts.

We may also miss a restart (read as non-NULL and became NULL=default).

In both cases, when the value sys->device_name is modified a restart will be
triggered anyway. So missing an extra one or adding one will make little
differences.