mmdevice: use a single string pointer for the device name
We only care about the name before getting sys->dev and we keep it
in case of restart on the same device.

We also need a way to tell if sys->dev has been acquired or if we failed
to do so. NULL alone in sys->dev can't tell that.

In OnDefaultDeviceChange() we don't care if the device was acquired or not.
We only care if we are using the default device when it changes.