qml: use player controller to probe meta data in `Player.qml`
Using playlist controller's current item is not a good idea
here because if the item is preparsed after it is read in
QML, the details are not updated since `PlaylistItem` is
a gadget and can not signal changes (and the item itself
does not change).

We can instead use player controller's meta data properties
directly, which makes more sense for the player view anyway.