qt: introduce `antialiasing` uniform variable in `SDFAARoundedTexture.frag`
In addition to the preprocessor definition, now there is an uniform variable
to dictate whether antialiasing should be active or not.

Research suggests that conditions based on uniform variables are optimized
well. Consumers may still remove the `ANTIALIASING` preprocessor define to
generate a shader that does not have the capability of applying AA. For
maintenance purposes, generating the shader without AA capability is left
for discussion until the build system supports generating shaders with
preprocessor defines.

Having a shader with no AA capability may get rid of few if statements,
though that would likely not impact the performance in any meaningful
way (see the note above). But at the same time, it would increase the
size, and possibly make antialiasing switch more laggy (though this
would not mean much, since antialiasing switch is not meant to be
animated, at least so far) due to shader switch even though shaders
may be cached, it can not be as fast as adjusting the uniform buffer.

It should be noted that we are already checking conditions based on
the uniform block in both `CROP_SUPPORT` and `BORDER_SUPPORT`
variations.